requirements.txt
apache-airflow
apache-airflow-providers-http
kafka-python
SQLAlchemy
apache-airflow
apache-airflow-providers-apache-spark
pyspark

